requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
streamlit>=1.35.0
